Extract public key and signature to new `sr25519` module in `subspace-core-primitives`
