+++ subspace/crates/pallet-subspace/src/mock.rs
@@ -18,7 +18,7 @@ use subspace_core_primitives::segments::{
     ArchivedBlockProgress, HistorySize, LastArchivedBlock, SegmentHeader, SegmentIndex, SegmentRoot,
 };
 use subspace_core_primitives::solutions::{Solution, SolutionRange};
-use subspace_core_primitives::PublicKey;
+use subspace_core_primitives::sr25519::PublicKey;
 use subspace_runtime_primitives::ConsensusEventSegmentSize;
 
 type Block = frame_system::mocking::MockBlock<Test>;
