+++ subspace/crates/sp-consensus-subspace/src/digests.rs
@@ -12,7 +12,8 @@ use sp_std::num::NonZeroU32;
 use subspace_core_primitives::hashes::Blake3Hash;
 use subspace_core_primitives::pot::PotOutput;
 use subspace_core_primitives::segments::{SegmentIndex, SegmentRoot};
-use subspace_core_primitives::solutions::{RewardSignature, Solution, SolutionRange};
+use subspace_core_primitives::solutions::{Solution, SolutionRange};
+use subspace_core_primitives::sr25519::RewardSignature;
 
 /// A Subspace pre-runtime digest. This contains all data required to validate a block and for the
 /// Subspace runtime module.
