+++ subspace/crates/subspace-farmer-components/benches/plotting.rs
@@ -6,7 +6,7 @@ use std::env;
 use std::num::NonZeroU64;
 use subspace_archiving::archiver::Archiver;
 use subspace_core_primitives::segments::{HistorySize, RecordedHistorySegment};
-use subspace_core_primitives::PublicKey;
+use subspace_core_primitives::sr25519::PublicKey;
 use subspace_farmer_components::plotting::{plot_sector, CpuRecordsEncoder, PlotSectorOptions};
 use subspace_farmer_components::sector::sector_size;
 use subspace_farmer_components::FarmerProtocolInfo;
