+++ subspace/crates/subspace-farmer-components/benches/reading.rs
@@ -12,7 +12,7 @@ use subspace_archiving::archiver::Archiver;
 use subspace_core_primitives::pieces::PieceOffset;
 use subspace_core_primitives::sectors::SectorId;
 use subspace_core_primitives::segments::{HistorySize, RecordedHistorySegment};
-use subspace_core_primitives::PublicKey;
+use subspace_core_primitives::sr25519::PublicKey;
 use subspace_farmer_components::file_ext::{FileExt, OpenOptionsExt};
 use subspace_farmer_components::plotting::{
     plot_sector, CpuRecordsEncoder, PlotSectorOptions, PlottedSector,
