+++ subspace/crates/subspace-farmer/src/plotter/cpu.rs
@@ -27,7 +27,7 @@ use std::sync::Arc;
 use std::task::Poll;
 use std::time::Instant;
 use subspace_core_primitives::sectors::SectorIndex;
-use subspace_core_primitives::PublicKey;
+use subspace_core_primitives::sr25519::PublicKey;
 use subspace_data_retrieval::piece_getter::PieceGetter;
 use subspace_farmer_components::plotting::{
     download_sector, encode_sector, write_sector, CpuRecordsEncoder, DownloadSectorOptions,
