+++ subspace/crates/subspace-node/src/chain_spec.rs
@@ -8,7 +8,7 @@ use std::marker::PhantomData;
 use std::num::NonZeroU32;
 use subspace_core_primitives::pot::PotKey;
 use subspace_core_primitives::solutions::SolutionRange;
-use subspace_core_primitives::PublicKey;
+use subspace_core_primitives::sr25519::PublicKey;
 use subspace_runtime::{
     AllowAuthoringBy, BalancesConfig, RuntimeConfigsConfig, RuntimeGenesisConfig, SubspaceConfig,
     SudoConfig, SystemConfig, WASM_BINARY,
