+++ subspace/crates/subspace-rpc-primitives/src/lib.rs
@@ -4,7 +4,8 @@ use parity_scale_codec::{Decode, Encode, EncodeLike, Input, Output};
 use serde::{Deserialize, Serialize};
 use std::time::Duration;
 use subspace_core_primitives::hashes::Blake3Hash;
-use subspace_core_primitives::solutions::{RewardSignature, Solution, SolutionRange};
+use subspace_core_primitives::solutions::{Solution, SolutionRange};
+use subspace_core_primitives::sr25519::RewardSignature;
 use subspace_core_primitives::SlotNumber;
 use subspace_farmer_components::FarmerProtocolInfo;
 use subspace_networking::libp2p::Multiaddr;
